requirements.txt
# Core FastAPI Framework
fastapi==0.124.0
uvicorn[standard]==0.32.1
python-multipart==0.0.20

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# Database
SQLAlchemy==2.0.44

# Rate Limiting
slowapi==0.1.9

# Pydantic for Data Validation
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5

# LangChain & AI
langchain==0.2.14
langchain-chroma==0.1.2
langchain-community==0.2.12
langchain-core==0.2.38
langchain-google-genai==1.0.10
langchain-openai==0.1.7
langchain-text-splitters==0.2.4
langchain-unstructured==0.1.2

# OpenAI API (for OpenRouter)
openai==1.109.1

# Embeddings
sentence-transformers==5.1.2
huggingface-hub==0.36.0

# Vector Store
chromadb==0.4.24

# Document Processing
unstructured==0.16.9
unstructured[pdf]==0.16.9
unstructured[docx]==0.16.9
pypdf==5.1.0
python-docx==1.1.2
Pillow==11.0.0
pdf2image==1.17.0

# Web Search
google-search-results==2.4.2

# HTTP Client
requests==2.32.3

# CORS Middleware (included in FastAPI but explicit)
starlette==0.41.3

# Additional Utilities
python-dotenv==1.0.1
typing_extensions==4.15.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.27.2

# Data Processing
numpy==1.26.4
pandas==2.2.3

# Token Management
tiktoken==0.8.0

# Async Support
aiofiles==24.1.0
anyio==4.7.0

# JSON Processing
orjson==3.10.12

# Logging
colorlog==6.9.0
